CRUD API States

+++ src/main/java/br/com/compasso/RestAPI/entity/State.java
@@ -27,5 +27,5 @@ public class State implements Serializable{
 
 	private String capital;
 	
-	private Integer area;
+	private Double area;
 }
